requirements.txt
dify_plugin
python-dotenv>=1.0.0
Pillow>=10.0.0